Update rad_haplotyper.pl
Hey Chris,

I'm not sure what's appropriate on github.  I'm experimenting, so feel free to ignore or delete this if you want.  

I commented above or on every line I edited with the letters CEB

I added an argument to rad_haplotyper that allows the user to specify  the path to the bam files.  I thought you might think that's useful.

The other argument I added, which may not be so useful but shouldn't mess anything up, enables radhaplotyper to be compatible with my modified dDocent file naming that include the assembly cutoffs.  Ex Pop1.3.6  would be pop1 mapped to an assembly that used cutoff1=3 and cutoff2=6.   This allows me to run dDocent simultaneously with different settings in the same directory.

Cheers,

Chris